Disabling dataloading prefetch to fix a bug.
There was a bug where consecutive training steps would use the same batch. The
number of training steps using the same batch is equal to `n_cpu`. This is to
do with the number of workers prefetching data in the dataloader, and somehow
each of them independently keeping track of their own `data_counter` member in
`mmap_batch_generator`. I don't completely understand how the prefetching
caused this, all I know is that it did.

The fix substantially changes the training dynamics by bringing stochasticity
back into training. In particular training error is much lower and validation
accuracy and recall are much higher. It's noteworthy that validation false
positives are also much higher. The training logic in this codebase is quite
idiosyncratic so it's likely some of the settings and logic need to be
recalibrated to work well with correct minibatch SGD to minimize false
positives.